Finished general/standard user gui functionality; you can now withdraw and show account details which finalizes the current 4 options the standard user has

+++ src/main/kotlin/org/setu/Screens.kt
@@ -136,7 +136,7 @@ class Screens {
             #         Welcome ${currentUser.firstName.replaceFirstChar { it.uppercase() }}        #
             ####################################
             How much would you like to Withdraw:     
-            [Current balance]:  ${currentUser.balance.toInt()}
+            [Current balance]:  ${currentUser.balance}
    
             Enter an amount: 
         """.trimIndent()
